+++ src/entity/relation.rs
@@ -1,6 +1,9 @@
-use crate::{EntityTrait, Identity, IdentityOf, Iterable, QuerySelect, Select};
+use crate::{unpack_table_ref, EntityTrait, Identity, IdentityOf, Iterable, QuerySelect, Select};
 use core::marker::PhantomData;
-use sea_query::{Alias, Condition, DynIden, JoinType, SeaRc, TableRef};
+use sea_query::{
+    Alias, Condition, DynIden, ForeignKeyCreateStatement, JoinType, SeaRc, TableForeignKey,
+    TableRef,
+};
 use std::fmt::Debug;
 
 /// Defines the type of relationship
@@ -306,6 +309,131 @@ where
             on_condition: b.on_condition,
             fk_name: b.fk_name,
         }
+    }
+}
+
+impl From<RelationDef> for ForeignKeyCreateStatement {
+    fn from(relation: RelationDef) -> Self {
+        let mut foreign_key_stmt = Self::new();
+        let from_tbl = unpack_table_ref(&relation.from_tbl);
+        let to_tbl = unpack_table_ref(&relation.to_tbl);
+        let from_cols: Vec<String> = match relation.from_col {
+            Identity::Unary(o1) => vec![o1],
+            Identity::Binary(o1, o2) => vec![o1, o2],
+            Identity::Ternary(o1, o2, o3) => vec![o1, o2, o3],
+        }
+        .into_iter()
+        .map(|col| {
+            let col_name = col.to_string();
+            foreign_key_stmt.from_col(col);
+            col_name
+        })
+        .collect();
+        match relation.to_col {
+            Identity::Unary(o1) => {
+                foreign_key_stmt.to_col(o1);
+            }
+            Identity::Binary(o1, o2) => {
+                foreign_key_stmt.to_col(o1);
+                foreign_key_stmt.to_col(o2);
+            }
+            Identity::Ternary(o1, o2, o3) => {
+                foreign_key_stmt.to_col(o1);
+                foreign_key_stmt.to_col(o2);
+                foreign_key_stmt.to_col(o3);
+            }
+        }
+        if let Some(action) = relation.on_delete {
+            foreign_key_stmt.on_delete(action);
+        }
+        if let Some(action) = relation.on_update {
+            foreign_key_stmt.on_update(action);
+        }
+        let name = if let Some(name) = relation.fk_name {
+            name
+        } else {
+            format!("fk-{}-{}", from_tbl.to_string(), from_cols.join("-"))
+        };
+        foreign_key_stmt
+            .name(&name)
+            .from_tbl(from_tbl)
+            .to_tbl(to_tbl)
+            .take()
+    }
+}
+
+/// Creates a column definition for example to update a table.
+/// ```
+/// use sea_query::{Alias, IntoIden, MysqlQueryBuilder, TableAlterStatement, TableRef};
+/// use sea_orm::{EnumIter, Iden, Identity, PrimaryKeyTrait, RelationDef, RelationTrait, RelationType};
+///
+/// let relation = RelationDef {
+///     rel_type: RelationType::HasOne,
+///     from_tbl: TableRef::Table(Alias::new("foo").into_iden()),
+///     to_tbl: TableRef::Table(Alias::new("bar").into_iden()),
+///     from_col: Identity::Unary(Alias::new("bar_id").into_iden()),
+///     to_col: Identity::Unary(Alias::new("bar_id").into_iden()),
+///     is_owner: false,
+///     on_delete: None,
+///     on_update: None,
+///     on_condition: None,
+///     fk_name: Some("foo-bar".to_string()),
+/// };
+///
+/// let mut alter_table = TableAlterStatement::new()
+///     .table(TableRef::Table(Alias::new("foo").into_iden()))
+///     .add_foreign_key(&mut relation.into()).take();
+/// assert_eq!(
+///     alter_table.to_string(MysqlQueryBuilder::default()),
+///     "ALTER TABLE `foo` ADD CONSTRAINT `foo-bar` FOREIGN KEY (`bar_id`) REFERENCES `bar` (`bar_id`)"
+/// );
+/// ```
+impl From<RelationDef> for TableForeignKey {
+    fn from(relation: RelationDef) -> Self {
+        let from_tbl = unpack_table_ref(&relation.from_tbl);
+        let mut foreign_key = Self::new()
+            .from_tbl(relation.from_tbl)
+            .to_tbl(relation.to_tbl)
+            .take();
+        let from_cols: Vec<String> = match relation.from_col {
+            Identity::Unary(o1) => vec![o1],
+            Identity::Binary(o1, o2) => vec![o1, o2],
+            Identity::Ternary(o1, o2, o3) => vec![o1, o2, o3],
+        }
+        .into_iter()
+        .map(|col| {
+            let col_name = col.to_string();
+            foreign_key.from_col(col);
+            col_name
+        })
+        .collect();
+        match relation.to_col {
+            Identity::Unary(o1) => {
+                foreign_key.to_col(o1);
+            }
+            Identity::Binary(o1, o2) => {
+                foreign_key.to_col(o1);
+                foreign_key.to_col(o2);
+            }
+            Identity::Ternary(o1, o2, o3) => {
+                foreign_key.to_col(o1);
+                foreign_key.to_col(o2);
+                foreign_key.to_col(o3);
+            }
+        }
+        if let Some(action) = relation.on_delete {
+            foreign_key.on_delete(action);
+        }
+        if let Some(action) = relation.on_update {
+            foreign_key.on_update(action);
+        }
+        let name = if let Some(name) = relation.fk_name {
+            name
+        } else {
+            format!("fk-{}-{}", from_tbl.to_string(), from_cols.join("-"))
+        };
+        foreign_key.name(&name);
+        foreign_key
     }
 }
